+++ examples/cpp/shared/example_vamana_with_compression.cpp
@@ -23,6 +23,7 @@
 #include "svs/orchestrators/dynamic_vamana.h"
 #include "svs/orchestrators/exhaustive.h"
 #include "svs/orchestrators/vamana.h"
+#include <cstdint>
 
 int main() {
     // STEP 1: Compress Data with LeanVec, reducing dimensionality to leanvec_dim dimensions
@@ -69,7 +70,7 @@ int main() {
     //! [Perform Queries]
 
     //! [Recall]
-    auto groundtruth = svs::load_data<int>(
+    auto groundtruth = svs::load_data<uint32_t>(
         std::filesystem::path(SVS_DATA_DIR) / "groundtruth_euclidean.ivecs"
     );
     double recall = svs::k_recall_at_n(groundtruth, results, n_neighbors, n_neighbors);